requirements.txt
einops==0.8.1
huggingface-hub==0.36.0
librosa==0.11.0
matplotlib==3.10.7
monai==1.5.1
monai-generative==0.2.3
numba==0.62.1
numpy==2.1.2
pandas==2.3.3
pytorch-lightning==2.5.6
scipy==1.15.3
soundfile==0.13.1
timm==1.0.22
torchinfo==1.8.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.1
